fix: removed extra navbar components
