--- src/components/NavBar-Orange/navbar.scss
@@ -1,28 +1,0 @@
-.NAVBAR {
-    background-color: #ffb181;
-    padding: 0.2% 0% 0.5% 0%;
-    align-items: flex-start;
-
-    .NAVTEXT {
-        font-size: 16px;
-        display: block;
-    }
-    .sticky {
-        position: fixed;
-        width: 100%;
-    }
-    .MLH-banner {
-        width: 170px;
-        height: 50px;
-        position: fixed;
-        margin: 0% 0% 0% .3%;
-    }
-
-    .MLH{
-        position: -webkit-sticky;
-        position: sticky;
-    }
-    .active:active {
-        background-color: white;
-    }
-}
