+++ src/components/NavBar/navbar.scss
@@ -1,6 +1,6 @@
 .NAVBAR {
     background-color: #ffb181;
-    padding: 0.5% 0% 0.5% 0%;
+    padding: 0.2% 0% 0.5% 0%;
     align-items: flex-start;
 
     .NAVTEXT {
@@ -13,10 +13,12 @@
     }
     .MLH-banner {
         width: 170px;
-        height: 40px;
+        height: 50px;
+        position: fixed;
+        margin: 0% 0% 0% .3%;
     }
 
-    .MLH {
+    .MLH{
         position: -webkit-sticky;
         position: sticky;
     }
